CTCTRADERS-2478 Refactor scheduler package
